requirements.txt
langgraph>=0.0.30
langchain>=0.1.10
langchain-groq>=0.1.0
fastapi>=0.110.0
uvicorn>=0.27.1
streamlit>=1.32.0
python-dotenv>=1.0.1
faiss-cpu>=1.8.0
sentence-transformers>=2.5.1
langchain-community>=0.0.25
langchain-huggingface>=0.0.1
pydantic>=2.6.4
